[blackout] use dev-topological for dev capability
In fxrev.dev/722416, the dev capability got renamed to dev-topological.
This was changed for the integration tests, but not the actual tests,
which means the actual tests are attempting to route a capability that
doesn't exist.

To try to prevent a change like this from not being propagated in the
future, organize the repeated set of cml fields into a blackout shard
that is included by the tests instead of repeating ourselves.

Change-Id: Ieb58c042d1ef854f20b46e58756e1aded6587f40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732602
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>