[wlantap] Remove ClearAssoc() MAC address NULL check
The MAC address is no longer verified in the mocked ClearAssoc()
function. In the implementation prior to the DFv2 migration in
fxrev.dev/660030, only the array pointer was checked to be non-null,
not the content of the MAC address itself.

Test: hw-sim test suite
Bug: 111067
Change-Id: I877efce7e37edfceb4723d54b4b1707582be3e4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734342
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>