[emu] Remove double-listing of PBMs on start
fxrev.dev/718948 introduced achange to product-bundle loading in the
emulator launcher that caused the list of available PBMs to be printed
twice, when a complete URL wasn't provided on the CLI. This is because
the product URL was needed in multiple locations, but the name was being
passed instead (which triggers resolution).

This CL adds a parameter to the select_product_bundle function,
should_print, which suppresses the output when set to false; this is true
in all cases except the known second resolution for `ffx emu start`.
This is considered an acceptable workaround, since the product-bundle
tool is deprecated pending PBv2 completion.

Bug: 110265
Change-Id: I18cb4f9e398836f89b4b8b431a41b02eb698d803
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731736
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>