[fidl][docs] Improve error example regex
This makes the regex more specific: rather than removing all comments,
it hones in on comments that are specifically likely to be copyrights.
Unfortunately, because the devsite regex engine does not allow multiline
move, we can't do more than that, like only removing the first 3 comment
lines in the file.

Test: manual
Bug: 108248
Change-Id: If5c42b11b7f96a0b4f8534f9d34a80d9e8c0e477
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732630
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>