[dfv2] Use node's real name for composite names
Node::name() returns the node's name with the ":" and "/"
replaced so it can be used as a component moniker.
Store the real name in the node so that we can use it
for composite devices correctly, so they show up with the
exact same name they were given in DFv1.

TEST: device-enumeration-test in next CL

Bug: 106517

Change-Id: I04d1a94916d050cbf73eed00af6b0efe6fc73078
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732875
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>