[vm][utest] Drop non-essential WaitForBlocked() from pager-writeback
Most of the tests in core-pager-writeback do not strictly need a call to
WaitForBlocked(), which can be flaky when the test is run as a component
and the test thread is blocked on a pager outside of the test. Remove
such instances of WaitForBlocked().

These tests would wait on a pager request immediately after the
WaitForBlocked() anyway, so the WaitForBlocked() was not strictly
required. The WaitForBlocked() check was not providing any meaningful
information in a component test.

Bug: 109851
Multiply: fuchsia-pkg://fuchsia.com/core-pager-writeback-test-package#meta/core-pager-writeback-test-package.cm
Change-Id: Ibebacd8355c37c7a6f13d4b6129247fbfa6a86d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733924
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>