[zxio] LRU eviction for dgram socket cache
Prevent unbounded growth in the client-side send cache by keeping a list
of the entries in most- to least-recently-used order. When an entry in
the cache is inserted or accessed, move it the front; when the cache
reaches its size limit, discard the least-recently-used entry.

Note that this CL also increases the size of the storage backing a
`zxio_t` in order to accommodate the `std::list` added to the route
cache in the client.

Fixed: 97260
Test: fx test zxio-unittest-package
      fx test netstack_with_fast_udp_dgramsocket_c_api_test
Change-Id: I6d780ed8d78135009ef93c20d61600685134c802
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729945
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>