[intel-i915-tgl] Add transcoder VRR reset step.
This CL implements step 2.h. in the section "Sequences for DisplayPort"
> "Disable Sequence" on pages 147-148 of IHD-OS-TGL-Vol 12-1.22-Rev2.0.

This CL does not fix the system hang in the attached bug. However, it
does bring our driver logic closer to the official documentation, so it
is a step in the right direction.

Bug: 111129
Change-Id: I39350d899b279fa38897d2202d7176d08524bbef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734782
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>