[dfv2] Include device-enumeration-test
Include the device-enumeration-test on DFv2 devices.
Update the test so it checks for the DFv2 composites.

This CL outright removes the check for:
pci-00:00.0-fidl

This is a composite that does not have a driver bound to it.
Composites without drivers don't show up in DFv2, and because
there's no driver bound to it, it's likely not worth checking
for in DFv1 either.

Bug: 106517

Change-Id: I9f9d73de3110501fb475c0a73a264f479878517c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731885
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>