[direct-mode] Optionally pass guest handle to process
This allows us to optionally pass the guest handle to a process launched
within direct-mode. This will be useful in an upcoming change, where we
can use the guest handle to control whether the process is allowed to
act like a kernel and create its own address spaces to manage.

This option is controlled through an environment variable, in order to
minimally interfere with the process being launched.

Bug: 95763
Change-Id: Ifd1093f80312b796843c0543ca80a6618a7e006a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734826
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>