[bt][intel] Update firmware reset after secure fw load
When loading firmware, we send a vendor reset with a boot address in the
parameters for it at the end.  This boot address was previously
hard-coded but is not static for all intel controllers, so instead read
the boot address from the firmware as we are sending it, and send the
correct address on reset.

Update some logging while we are here (newlines are not required, add
more info to some)

Bug: 108011
Test: boot atlas and observe info in logs from firmware
Change-Id: Iea2f4cecda6fcb55f3e6510cabd911a50de47b48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725383
Reviewed-by: Ben Lawson <benlawson@google.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>