[virtualization] Add integration test for restarting a guest
This test validates that a guest can be cleanly restarted (including
bringing up all of the devices, etc.), without restarting the full
component.

Since this test is polling guest manager for when the VMM has stopped,
this needs to fix a minor race where the VMM component only reports that
it's stopped after it destroys the VMM object.

Note that this currently skips termina as termina requires that the
guest manager is torn down since it requires /dev, and drops /dev when
starting the VMM for security.

Bug: 104989
Multiply: virtualization_shutdown_tests
Change-Id: Ibc195d35e4f1e52efc1ac8fdba1b0da1106a3059
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728225
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>