[pbm] Always default to the local bundle in-tree
This CL modifies the product-bundle selection code (called when
disambiguating multiple available product bundles) to always default to
the bundle generated by the in-tree build when it's available. If no
local build is available (which shouldn't be common), it will fall back
on the previous behavior of selecting the first in a reverse-sorted
list.

Bug: 98978
Change-Id: I4e16231497926aa5f4e31fc0db52a32f6853c23b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734987
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>