[driver_runtime] Fix race condition in async dispatcher
CancelWait() could fail in a synchronized dispatcher because the
underlying async loop is multithreaded - if the wait was signaled and
another thread pulled the packet off the port, but the callback was not
yet invoked, then CancelWait() would return ZX_ERR_NOT_FOUND and the
callback would later be invoked. This CL makes it so that the dispatcher
will cancel the wait and return ZX_OK in this case.

Fixed: 109988
Multiply: fuchsia-pkg://fuchsia.com/driver_runtime_tests#meta/driver_runtime_tests.cm : 100
Change-Id: Ib93d595b9aa711b1d0b25ebbb57deb54c5d6ccb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728950
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Ian Fisher <iafisher@google.com>