[text-input] An attempt to resolve a test flake
This test was flaking when many repeated taps on the browser screen did not result in a focus change
on the input field.

chaopeng@ and I think that this might happen as a result of (1) the browser window gaining focus before JS is fully loaded (which is corroborated by logs) and (2) the 'autofocus' attribute on the text
field, which means the field will be focused when the page loads. (we couldn't corroborate this with jsfiddle for example, as 'autofocus' seems to behave differently there)

So, if (1) happens, then (2) may not result in a focus change notification, because JS didn't install the focus handler yet when the focus was changed,. In addition, autofocus would have already placed the focus on the text field by the time JS is operational. This would lead to the code never getting notified
of focus gain.

If this hypothesis iscorrect, then the removal of the "autofocus" attribute might resolve the 
issue, modulo any misunderstandings of the underlying mechanism on my behalf.

Bug: 111219
Change-Id: I6886446a674f06f331d7dc970f5e2a900073578f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735306
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>