[ffx][flash][test] Make test harness more robust
Removes the public "variables" member of FakeServiceCommands, and
replaces it with `HashMap<String, (String, u32)>`. Also adds functions
to set the variable and it's associated value, as well as retrieve the
number of times a particular variable has been called.

This removes the need for developers to care about managing the order by
which the variables are retrieved from the device in their test, and
also allows them to make assertions on the number of times a variable
has been retrieved by the calling code.

Change-Id: Ib0916bae1d6ed867159192399f48b5bd996342e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734986
Commit-Queue: Colin Nelson <colnnelson@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>