[a11y] InjectPointerEventAction: don't crash on null node
Early return instead of crashing when GetSemanticNode returns nullptr.

This is in response to an issue we were seeing in production:
fxbug.dev/110562
The issue likely still exists, but now we'll merely fail to process the
pass-through gesture (instead of crashing).

We've added an error log, so we can track this down in the future.

Bug: 110562
Test: a11y-lib-tests
Change-Id: I94c962eb8253114a90d3bda2bce778264631d93a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733659
Reviewed-by: Alexander Brusher <abrusher@google.com>
Commit-Queue: Kevan Hollbach <khollbach@google.com>