[build] Fix clang's ubsan-trap config.
The previous configuration for ubsan-trap was not enabling the
sanitizer.

According to the clang documentation[0], specifying `-fsanitize-trap`
does not enable the corresponding sanitizer. `-fsanitize` is also
needed.

This change adds a `-fsanitize=undefined` to the ubsan-trap config,
enabling ubsan traps in non-gcc based builds.

This commit also adds a definition for std::nothrow to the EFI
allocator.  Adding the ubsan instrumentation caused the thinlto builder
to fail.

[0]: https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html

Change-Id: Ie94f0c360bbdf78761937888b2b64bcb0f148783
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720702
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>