[fidl][cpp] Flexible enums default initialize to Unknown
This is a prospective API change to be discussed during the
API review of the new C++ bindings.

It's more difficult to disallow default construction
of flexible enums because they are used in wire types which overall
is reliant on default construction.

Between zero or Unknown, Unknown is a slightly better choice because
it wouldn't accidentally send whatever element that happen to be
defined to zero.

RELNOTES: When default constructing flexible enums, note that they will
hold an unknown value instead of zero. Code that previously send default
constructed enums down the wire may now send a unknown value instead of
the member whose ABI is zero. To avoid sending unknowns, make sure
flexible enums are all properly initialized to a desired value instead
of relying on C++ default initialization.

Change-Id: I58246f16b532440c09c228cbe195d4e720a50210
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726710
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>