[component_manager] align resolution Context with pkg ResolutionContext
When the PackageResolver API was updated to support
ResolutionWithContext, the FIDL type declaration for the
ResolutionContext used a struct, with field `bytes`. This was an
improvement over the Component resolution::Context type that used an
alias for a byte array, because language bindings for the alias type
implemented the type as an array of bytes. The struct type in
PackageResolver is more type safe and more readable since the language
bindings always refer to the type as `ResolutionContext`.

This CL aligns the Component FIDL to use a similar struct declaration,
and updates API usage (such as in component_manager) accordingly.

Bug: 100050

Fixed: 104038

Change-Id: Ifd3f4d6bdcae40a7b4b26243ea6c1871d7e5d70e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735405
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
API-Review: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Rich Kadel <richkadel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>