[power][test] Ignore FIDL result in mock
Changes MockActivityService to ignore the result of calling
`on_state_changed` on the activity listener channel. Previously, an
error would cause a crash, which is not the correct behavior because the
service should be resilient to calling a method on a closed channel.

Bug: 104266
Test: fx test //src/power/power-manager/integration
Change-Id: I454790ee80fe977d890274656149d5e4a7901f4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735765
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Patrick Shickel <pshickel@google.com>
Reviewed-by: Didi She <didis@google.com>