[sdmmc][fidl] Add method to re-enable the in-band interrupt
Add a method to ack the SDIO in-band interrupt to allow it to be
delivered again. SDMMC drivers will be changed to disable the interrupt
immediately after it is received, and only re-enable it after this
method has been called. The in-band interrupt is asserted by the card
until all interrupt sources have been cleared by function drivers, so
disabling it at the controller is needed to prevent a flood of
interrupts.

Bug: 110355
Test: CQ
Change-Id: I13d6e11f54f379d993bcf73354437a0411bd5d0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729783
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
API-Review: Christopher Anderson <cja@google.com>