[sdio][core] Keep in-band interrupt disabled until a client handles it
An upcoming change will make SDMMC drivers disable the in-band interrupt
and leave it up to the SDIO core driver to re-enable it. Clearing the
interrupt source on the card is function-specific, so the best the core
driver can do is to re-enable the interrupt after a function driver
notifies it that the last one has been handled.

Bug: 110355
Test: fx test sdmmc-test
Change-Id: I1b22bc2954024d32a7c516bc42c96e2371bd633e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729903
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>