[lowpan][thread-config-manager] Don't "leak" memory on every read.
As seen in <fxbug.dev/109073>, the size of our `lowpan-ot-driver`
process was increasing in a slow but unbounded fashion. After some
investigation, it was determined that the the behavior of
`rapidjson::MemoryPoolAllocator` was most likely responsible, which
explicitly does not ever free allocated memory blocks[1][].

We believe the long term fix is to re-engineer this code entirely to
avoid using the DOM interface. However, that would be a large and
potentially destabilizing change.

This change simply avoids a copy operation, which was determined to be
the source of almost all of our increasing memory footprint. While
it does not fix the memory leak entirely, it does reduce it from
around 500kb/day to something closer to 30kb/day or less.

Additional investigation is warranted, but this change seems like a
reasonable first step.

[1]: https://rapidjson.org/classrapidjson_1_1_memory_pool_allocator.html

Bug: 109073
Test: Verified manually with ASAN enabled
Change-Id: I3030161de1b7a402868a5855a846f2363e16989f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735715
Commit-Queue: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Parag Dixit <paragdixit@google.com>
(cherry picked from commit d2abf9c3fe0969eeaab93cc7422f045f10c27d0a)