[identity] Update all component log tags.
With the old sytle of diagnostics init each log statement included the
component in addition to the domain but recently we've lost that tag
which makes some of the logs hard to follow.

This CL adds the component name as a tag on all logs and also moves from
the ~2017 "auth" nomenclature to the "identity" we use in interfaces
and the source tree.

Change-Id: Ib01c2c3787de16aa3b7ca162ebed666c2c22d5d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735406
Reviewed-by: Angad Pal Singh <apsbhatia@google.com>
Reviewed-by: Ben Wright <benwright@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>