[input] Starts reporting AltGr as a modifier
The AltGr key is special because it is not defined in the USB HID.
So the only way an AltGr modifier can be reported is by examining
the KeyMeaning, which is not how our current modifier tracker
works.

For this reason, another modifier handler is required, which
fixes up the modifiers reported with the modifier bits
derived from key meanings.

At the moment, only AltGr is handled, since this was
immediately needed for Chromium.

Bug: 110071, 109987
Change-Id: I31748c8420e5086e1f09436ab7949e28f61fcf97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728742
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>