[subprocess] Fix race in test.
One of the tests attempts to kill a subprocess that doesn't finish on
its own even after receiving a SIGTERM. It does this by having the
subprocess execute a `sleep` during which it wouldn't respond to the
SIGTERM so that we can execute the path that waits for a grace period to
let the subprocess clean up and then sends a SIGKILL when the subprocess
fails to clean up in time. In the very rare occasion, this test would
flake because we'd cancel the context and send the SIGTERM before the
subprocess script runs `sleep` and it would just exit and clean up
normally. This change pgreps for the sleep process to make sure it was
run before we send the SIGTERM.

Test: Ran `fx go test -failfast -count=10000 ./tools/lib/subprocess`
twice.

Bug: 111211
Change-Id: I9ed10c2eba66be376c0f106f672d073e9a15a3a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734491
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>