[pb] Introduce pbtool which spits out a build archive
In order to facilitate the infra transition from build archives to
product bundles, we are introducing this tool which can take a product
bundle and spit out a build archive. This tool is only assumed to be
used by infra, therefore we do not want to add this to ffx. We are
writing this in Rust in order to share a lot of existing code with ffx.

Bug: 110791
Change-Id: Icf4b7f320db97637a2f56b3cd4b0950b4690fcaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733942
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>