[debugger] Replace the "actions" system.
The actions system runs a sequence of commands. It is used for the
started attach/connect switches, and also a custom file that users can
provide to run some startup commands.

The way the actions work is that it's difficult to convert to the new
CommandContext system, especially the tests. It's also relatively
complex for what it does.

This patch replaces it with a simple function that chains the callbacks.

Change-Id: I3924f2cbeaf14875ba0faf0ffdfe0edffe8e26ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730005
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>