[nxpfmac] Ack in-band interrupt
Recent changes to the SDIO fidl protocol added a function to ack in-band
interrupts. Without calling the ack function the IRQ loop would get
stuck because every interrupt after the first one would be masked.

Fix this by calling the new function to ack the interrupt each time
we're going to wait for it.

Test: Manual testing, fx test nxpfmac-sdio-tests
Change-Id: I81967b51f5cd8d3be0030359ab9a3b14d69c9a73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736184
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>