[pbm] Prevent overwriting package repos on get
The attached bug describes how a package repo can be overwritten by
downloading another product bundle with the same short-name, since that
short-name is the default repo name for a new bundle. This CL prevents
downloading and adding a repo during product-bundle get if one already
exists with the indicated name. The error message instructs the user to
specify a new name with the --repository flag.

Bug: 104925
Change-Id: I2004595250519e62bc7873e99057b54c49f2f55c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735685
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>