[netstack3] `shutdown` and `close` for TCP core sockets
Similar to posix `shutdown` and `close`: `shutdown` queues a FIN for a
connection; `close` first does `shutdown`, but also removes the socket
from the socket map if the connection becomes `Closed`.

Note that SHUT_RD is not implemented in Core, because it is not visible
from the TCP state machine itself, and usually have divergent behaviors
on different platforms, we will handle it in Bindings.

Test: fx test netstack3-core-test-package

Fixed: 110573
Bug: 98970, 103979
Change-Id: I13430fb404eeaac7f736647113432a83351358eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735285
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>