[wlan][policy] Don't expose SME for scanning.
When scanning or during network selection, the IfaceManager
hands out a raw client SME for other clients to use to
perform scans.  While the SME is intended to be used only
for scanning, it could be used to change the client
connectivity state as well.

This change wraps the SME in a struct that only allows for
the scan command to be passed through to the SME.  It also
allows us to store some additional metadata about the
client interface associated with the SME (namely the ID)
to be used in the future for reporting events from which we
would like to recover.

Bug: 60089
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: I9e09d9607526b2a46caf3932577dcfb3f7dbc6aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734668
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>