[audio] Producer nodes can have at most one output edge
Originally, the mixer graph allowed each producer node to have multiple
output edges. This is functionally unnecessary (you can fan-out any node
by placing a splitter after that node), but it's more efficient to
implement fan-out directly on a producer node: the splitter requires a
copy, while in a producer, each output edge can read directly from the
producer's data source without a copy.

However, we discovered a challenging interaction with Start and Stop
commands. The new versions of these commands, proposed for the v2 media
APIs, have the following semantics:

- Start and Stop commands can be scheduled to happen in the future.
  There is at most one outstanding Start or Stop command. If a second
  command arrives before the first completes, the first command is
  cancelled and replaced by the second command.

Suppose a producer node has two outgoing edges, each connected to a
different output device, where the two edges have lead times 100ms and
50ms, respectively. Internally, we create one ProducerStage for each
output edge. Although these ProducerStages read from the same data
source, the run independently, and because of the different lead times,
ProducerStage P1 runs 50ms ahead of ProducerStage P2.

Each incoming Start or Stop command must be applied to both P1 and P2,
or cancelled and not applied to either P1 or P2. For each command, we
must decide whether to apply or cancel the command before the command
would be applied by P1, and if the command is applied, we must hold onto
this decision until the command is applied by P2. There is no limit on
the number of commands that can arrive in those intervening 50ms (and
there are realistic scenarios that can result in spam, such as a user
spamming the mouse to play/pause/play/pause/etc). This implies a command
buffer with unbounded length.

While there are ways to give this buffer a bounded length, there's no
way to keep P1 and P2 in sync without adding a command buffer to P2.

We decided it would be simpler to forbid multiple output edges at a
producer node. If the client wants to fan-out from a producer, they must
insert a splitter immediately after the producer. This makes Start and
Stop commands much easier to implement -- we never need to buffer more
than one pending command.

This CL removes the ability to add multiple outgoing edges to a producer
node. The FIDL comments have been updated and MetaProducerNode (the
container for a producer with multiple output edges) has been deleted.
StreamSinkServer is also simplified.

Bug: 87651
Change-Id: Ie6bfa40c649a740d13ad8f04a194c5e3210e8ae7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735266
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
API-Review: Dale Sather <dalesat@google.com>