[devfs] Reimplement using fs::Vnode
Prior to this change devfs did not properly enforce hierarchical rights;
to allow clients that rely on this omission to continue functioning, a
virtual method is added to `fs::Vnode` that allows rights checking to be
skipped to avoid breaking clients which rely on this absence of rights
checking.

This also fixes the behavior of `fuchsia.io/Directory.Clone` in the
in-tree C++ VFS library by appending `fuchsia.io/OpenFlags.DIRECTORY`.
This requires a new compliance test which is not included here; its
creation is tracked in https://fxbug.dev/111302.

Fixed: 38246
Bug: 101092
Bug: 110485
Bug: 110566
Bug: 111253
Bug: 111302
Change-Id: Ib4375c15263cee81e7e91467b5038cc1ce714fdd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729224
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>