[fidl][example-tester] Support golden testing
Instead of asserting on logs one by one, which produces very verbose
tests, the new `assert_logq_eq_to_golden` function does exactly what it
says on the tin: compare all of the logs for some component to a golden
string, imported into the test runner as a GN `resource`.

There are two things lacking to this approach. First, it would be nice
to have something like `fx regen-goldens`, though as far as I know
something like that would only work for host-side tests. As things
stand, users have to copy-paste golden output manually. I've tried to
add some niceties in terms of terminal-output to make this a bit easier,
but it is still rather annoying.

The second issue is that there is no way to filter logs, which will be
added in a future CL (ex: `assert_filtered_logs_eq_to_golden` or
similar). The reason this will be nice is that certain runtimes (like
the Dart VM) produce their own logs, so to ensure we compare the same
set of logs for each language, we'll probably need some filtering
capabilities.

All existing tests (key-value-store baseline and addreaditem cases) have
been migrated to use goldens.

Test: fx test examples_fidl example_tester
Bug: 109319
Change-Id: I8dd78519ad5f3b599ea5b208689cbf2bd1611c0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736084
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>