[storage-manager] AccountManager tests should generate StorageManager objects by factory.
A future change will expect AccountManager to accept a lambda which can
generate a struct which implements StorageManager, rather than a single
struct which implements StorageManager. This CL goes through all of the
testing callsites to perform that refactor eagerly, so that review of
that change is more self-contained.

Change-Id: Ic2c41f3f85a52f3c58b2e246f4c97071935ad95d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735842
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: James Buckland <jbuckland@google.com>