[spi] Add SPI Banjo device
Similar to fxr/713442, the old arrangement where the SPI device is both
bindable and adds a FIDL child to itself doesn't work in DFv2. So
instead we make a separate Banjo device that is the child of the SPI
device and the sibling of the FIDL device.

Test: booted Chromebook on DFv2 and verified that SPI devices now work
Fixed: 108557
Bug: 103068
Change-Id: I67eee93bebce0795feb6a4e1723bc78204ca1b60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736123
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Ian Fisher <iafisher@google.com>